+++ src/store/moderation/moderationActions.js
@@ -1,51 +1,58 @@
 import moderationService from '../../services/moderationService';
 
 export default {
+  reset,
   fetchNonModeratedScreenshots,
   fetchApprovedScreenshots,
   fetchRejectedScreenshots,
   fetchModeratedByYouScreenshots,
 };
 
-function fetchNonModeratedScreenshots() {
+function reset() {
+  return { type: 'MODERATION-RESET' };
+}
+
+function fetchNonModeratedScreenshots(params) {
   return async dispatch => {
     dispatch({ type: 'MODERATION-LOADING' });
-    const payload = await moderationService.fetchNonModerated();
+    const payload = await moderationService.fetchNonModerated(params);
     dispatch({
-      type: 'MODERATION-WAITING-LOADED',
+      type: 'MODERATION-LOADED',
       payload,
     });
   };
 }
 
-function fetchApprovedScreenshots() {
+function fetchApprovedScreenshots(params) {
   return async dispatch => {
     dispatch({ type: 'MODERATION-LOADING' });
-    const payload = await moderationService.fetchApproved();
+    const payload = await moderationService.fetchApproved(params);
     dispatch({
-      type: 'MODERATION-APPROVED-LOADED',
+      type: 'MODERATION-LOADED',
       payload,
     });
   };
 }
 
-function fetchRejectedScreenshots() {
+function fetchRejectedScreenshots(params) {
   return async dispatch => {
     dispatch({ type: 'MODERATION-LOADING' });
-    const payload = await moderationService.fetchRejected();
+    const payload = await moderationService.fetchRejected(params);
     dispatch({
-      type: 'MODERATION-REJECTED-LOADED',
+      type: 'MODERATION-LOADED',
       payload,
     });
   };
 }
 
-function fetchModeratedByYouScreenshots() {
+function fetchModeratedByYouScreenshots(params) {
   return async dispatch => {
     dispatch({ type: 'MODERATION-LOADING' });
-    const nonModeratedScreenshots = await moderationService.fetchModeratedByYou();
+    const nonModeratedScreenshots = await moderationService.fetchModeratedByYou(
+      params
+    );
     dispatch({
-      type: 'MODERATION-BY_YOU-LOADED',
+      type: 'MODERATION-LOADED',
       payload: nonModeratedScreenshots,
     });
   };